refactor(platform/api): extract tokens and project service into domain
