+++ platform/api/tests/test_ws_serve.py
@@ -35,10 +35,10 @@ from starlette.websockets import WebSocketDisconnect
 
 from hexgate_api import main
 from hexgate_api.main import app
+from hexgate_api.domains.tokens.service import mint_dev_token
 from hexgate_api.services import (
     DEFAULT_PROJECT_ID,
     ensure_default_project,
-    mint_dev_token,
 )
